1. 게시글 삭제 안되는 문제 수정 2. 몇몇 빌드 타입 환경변수를 런타임으로 변경 3. oauth logo 추가

+++ server/src/main/java/com/anamensis/server/provider/AwsS3Provider.java
@@ -176,12 +176,20 @@ public class AwsS3Provider {
     }
 
     public Mono<Void> deleteS3(String filePath, String filename) {
-        s3Client.deleteObject(
-            DeleteObjectRequest.builder()
-                .bucket(bucket)
-                .key(filePath.substring(1) + filename)
-                .build()
-        );
+
+        if(ACTIVE_S3) {
+            s3Client.deleteObject(
+                DeleteObjectRequest.builder()
+                    .bucket(bucket)
+                    .key(filePath.substring(1) + filename)
+                    .build()
+            );
+            return Mono.empty();
+        }
+
+        File file = new File(FILE_STORAGE_DIR + filePath + filename);
+        file.delete();
+
         return Mono.empty();
     }
